try to isolate anything specific to a hardware platform in hw.py

+++ logsupport.py
@@ -7,11 +7,14 @@ wc = webcolors.name_to_rgb
 import time
 import os
 import re
+from hw import disklogging
 
 Info = 0
 Warning = 1
 Error = 2
 
+
+# disklogging = True
 
 class Logs:
     livelog = True
@@ -23,18 +26,19 @@ class Logs:
     def __init__(self, screen, dirnm):
         self.screen = screen
         self.logfontsize = 23  # todo pixel
-        cwd = os.getcwd()
-        os.chdir(dirnm)
-        q = [k for k in os.listdir('.') if 'Console.log' in k]
-        if "Console.log." + str(config.maxlog) in q:
-            os.remove('Console.log.' + str(config.maxlog))
-        for i in range(config.maxlog - 1, 0, -1):
-            if "Console.log." + str(i) in q:
-                os.rename('Console.log.' + str(i), "Console.log." + str(i + 1))
-        os.rename('Console.log', 'Console.log.1')
-        self.disklogfile = open('Console.log', 'w')
-        os.chmod('Console.log', 0o555)
-        os.chdir(cwd)
+        if disklogging:
+            cwd = os.getcwd()
+            os.chdir(dirnm)
+            q = [k for k in os.listdir('.') if 'Console.log' in k]
+            if "Console.log." + str(config.maxlog) in q:
+                os.remove('Console.log.' + str(config.maxlog))
+            for i in range(config.maxlog - 1, 0, -1):
+                if "Console.log." + str(i) in q:
+                    os.rename('Console.log.' + str(i), "Console.log." + str(i + 1))
+            os.rename('Console.log', 'Console.log.1')
+            self.disklogfile = open('Console.log', 'w')
+            os.chmod('Console.log', 0o555)
+            os.chdir(cwd)
 
     def Log(self, entry, severity=Info, diskonly=False):
         """
@@ -44,11 +48,12 @@ class Logs:
         """
         if not diskonly:
             self.log.append((severity, entry))
-        self.disklogfile.write(time.strftime('%H:%M:%S')
-                               + ' Sev: ' + str(severity) + " " + entry.encode('ascii',
+        if disklogging:
+            self.disklogfile.write(time.strftime('%H:%M:%S')
+                                   + ' Sev: ' + str(severity) + " " + entry.encode('ascii',
                                                                                errors='backslashreplace') + '\n')
-        self.disklogfile.flush()
-        os.fsync(self.disklogfile.fileno())
+            self.disklogfile.flush()
+            os.fsync(self.disklogfile.fileno())
         if self.livelog and not diskonly:
             if self.livelogpos == 0:
                 config.screen.fill(wc('royalblue'))
